requirements.txt
streamlit
pandas
numpy
matplotlib
reportlab
openai
python-dotenv
openpyxl
